refactor: Do not use default export & separate files for types
