+++ src/index.ts
@@ -1,14 +1,14 @@
 export * from './useClickOutside';
-export { default as useCurrentDate } from './useCurrentDate';
-export { default as useDocumentEvent } from './useDocumentEvent';
-export { default as useFullState } from './useFullState';
-export { default as useGlobalState } from './useGlobalState';
-export { default as useIdle } from './useIdle';
-export { default as useInterval } from './useInterval';
-export { default as useKeyboardEvents } from './useKeyboardEvents';
-export { default as useNativeEvent } from './useNativeEvent';
-export { default as useWindowResize } from './useWindowResize';
-export { default as useScroll } from './useScroll';
-export { default as useTimeout } from './useTimeout';
-export { default as useToggle } from './useToggle';
-export { default as useWindowEvent } from './useWindowEvent';
+export * from './useCurrentDate';
+export * from './useDocumentEvent';
+export * from './useFullState';
+export * from './useGlobalState';
+export * from './useIdle';
+export * from './useInterval';
+export * from './useKeyboardEvents';
+export * from './useNativeEvent';
+export * from './useWindowResize';
+export * from './useScroll';
+export * from './useTimeout';
+export * from './useToggle';
+export * from './useWindowEvent';
