+++ src/useDocumentEvent/useDocumentEvent.ts
@@ -1,13 +1,16 @@
-import useNativeEvent from '../useNativeEvent';
-import { UseDocumentEvent } from './typings';
+import { useNativeEvent, UseNativeEventOptions } from '../useNativeEvent';
+
+export type UseDocumentEvent = <K extends keyof DocumentEventMap>(
+  type: K,
+  listener: EventListener,
+  options?: boolean | UseNativeEventOptions,
+) => void;
 
 /**
  * Simplifies the subscribing to events on document. Deletes the subscription after unmount component
  *
  * @see https://kundinos.github.io/docs/docs/react-hooks/hooks/use-document-event
  */
-const useDocumentEvent: UseDocumentEvent = (type, listener, options) => {
+export const useDocumentEvent: UseDocumentEvent = (type, listener, options) => {
   return useNativeEvent(document, type, listener, options);
 };
-
-export default useDocumentEvent;
