--- src/useFullState/typings.ts
@@ -1,3 +1,0 @@
-import { Dispatch, SetStateAction } from 'react';
-
-export type UseFullState = <S>(initialState?: S | (() => S)) => [S, Dispatch<SetStateAction<S>>, S];
