--- src/useIdle/typings.ts
@@ -1,7 +1,0 @@
-export interface UseIdleOptions {
-  timeout?: number;
-  onIdle?: (e?: Event) => void;
-  onWakeUp?: (e: Event) => void;
-}
-
-export type UseIdle = (options?: UseIdleOptions) => boolean;
