+++ src/useInterval/useInterval.test.ts
@@ -1,6 +1,6 @@
 import { renderHook } from '@testing-library/react-hooks';
 
-import useInterval from './useInterval';
+import { useInterval } from './useInterval';
 
 jest.useFakeTimers();
