+++ src/useNativeEvent/index.ts
@@ -1,2 +1,1 @@
-export * from './typings';
-export { default } from './useNativeEvent';
+export * from './useNativeEvent';
