+++ src/useNativeEvent/useNativeEvent.ts
@@ -1,13 +1,24 @@
-import { MutableRefObject, useEffect } from 'react';
+import { MutableRefObject, RefObject, useEffect } from 'react';
 
-import { UseNativeEvent } from './typings';
+export type Target<T extends EventTarget> = T | RefObject<T> | MutableRefObject<EventTarget>;
+
+export interface UseNativeEventOptions extends AddEventListenerOptions {
+  initial?: boolean;
+}
+
+export type UseNativeEvent = (
+  target: Target<EventTarget>,
+  type: string,
+  listener: EventListener,
+  options?: boolean | UseNativeEventOptions,
+) => void;
 
 /**
  * Simplifies the subscribing to events on any HTML element or React ref. Deletes the subscription after unmount component
  *
  * @see https://kundinos.github.io/docs/docs/react-hooks/hooks/use-native-event
  */
-const useNativeEvent: UseNativeEvent = (target, type, listener, options) => {
+export const useNativeEvent: UseNativeEvent = (target, type, listener, options) => {
   const isRef = !!(target as MutableRefObject<HTMLElement>).current;
   const initTarget = isRef
     ? (target as MutableRefObject<HTMLElement>).current
@@ -23,5 +34,3 @@ const useNativeEvent: UseNativeEvent = (target, type, listener, options) => {
     };
   }, [initTarget, listener, options, type]);
 };
-
-export default useNativeEvent;
