+++ src/useTimeout/index.ts
@@ -1,2 +1,1 @@
-export * from './typings';
-export { default } from './useTimeout';
+export * from './useTimeout';
