+++ src/useTimeout/useTimeout.ts
@@ -1,6 +1,20 @@
 import { useCallback, useEffect, useRef } from 'react';
 
-import { UseTimeout, Cleanup } from './typings';
+export type UseTimeoutCleanup = void | (() => void);
+
+export type UseTimeoutCallback = () => UseTimeoutCleanup;
+
+export type UseTimeoutDelay = number;
+
+export interface UseTimeoutResult {
+  reset: () => void;
+  repeat: () => void;
+}
+
+export type UseTimeout = (
+  callback: UseTimeoutCallback,
+  delay?: UseTimeoutDelay,
+) => UseTimeoutResult;
 
 /**
  * Declarative version of setTimeout
@@ -9,8 +23,8 @@ import { UseTimeout, Cleanup } from './typings';
  *
  * @see https://kundinos.github.io/docs/docs/react-hooks/hooks/use-timeout
  */
-const useTimeout: UseTimeout = (callback, delay) => {
-  const refCleanup = useRef<Cleanup>(null);
+export const useTimeout: UseTimeout = (callback, delay) => {
+  const refCleanup = useRef<UseTimeoutCleanup>(null);
   const refTimeoutId = useRef<NodeJS.Timeout>();
   const refIsCalled = useRef(false);
 
@@ -45,5 +59,3 @@ const useTimeout: UseTimeout = (callback, delay) => {
 
   return { reset, repeat };
 };
-
-export default useTimeout;
