--- src/useToggle/typings.ts
@@ -1,1 +1,0 @@
-export type UseToggle = (initialState?: boolean | (() => boolean)) => [boolean, () => void];
