+++ src/useToggle/useToggle.test.ts
@@ -1,6 +1,6 @@
 import { act, renderHook } from '@testing-library/react-hooks';
 
-import useToggle from './useToggle';
+import { useToggle } from './useToggle';
 
 describe('Main behavior', () => {
   test('Should be defined', () => {
