--- src/useWindowEvent/typings.ts
@@ -1,7 +1,0 @@
-import { UseNativeEventOptions } from '../useNativeEvent';
-
-export type UseWindowEvent = <K extends keyof WindowEventMap>(
-  type: K,
-  listener: EventListener,
-  options?: boolean | UseNativeEventOptions,
-) => void;
