+++ src/useWindowEvent/useWindowEvent.ts
@@ -1,13 +1,16 @@
-import useNativeEvent from '../useNativeEvent';
-import { UseWindowEvent } from './typings';
+import { useNativeEvent, UseNativeEventOptions } from '../useNativeEvent';
+
+export type UseWindowEvent = <K extends keyof WindowEventMap>(
+  type: K,
+  listener: EventListener,
+  options?: boolean | UseNativeEventOptions,
+) => void;
 
 /**
  * Simplifies the subscribing to events on window. Deletes the subscription after unmount component
  *
  * @see https://kundinos.github.io/docs/docs/react-hooks/hooks/use-window-event
  */
-const useWindowEvent: UseWindowEvent = (type, listener, options) => {
+export const useWindowEvent: UseWindowEvent = (type, listener, options) => {
   return useNativeEvent(window, type, listener, options);
 };
-
-export default useWindowEvent;
