--- src/useWindowResize/typings.ts
@@ -1,6 +1,0 @@
-import { UseNativeEventOptions } from '../useNativeEvent';
-
-export type UseWindowResize = (
-  listener: EventListener,
-  options?: boolean | UseNativeEventOptions,
-) => void;
