+++ src/useWindowResize/useWindowResize.ts
@@ -1,13 +1,16 @@
-import useWindowEvent from '../useWindowEvent';
-import { UseWindowResize } from './typings';
+import { useWindowEvent } from '../useWindowEvent';
+import { UseNativeEventOptions } from '../useNativeEvent';
+
+export type UseWindowResize = (
+  listener: EventListener,
+  options?: boolean | UseNativeEventOptions,
+) => void;
 
 /**
  * Simplifies the subscribing to resize of window. Deletes the subscription after unmount component
  *
  * @see https://kundinos.github.io/docs/docs/react-hooks/hooks/use-window-resize
  */
-const useWindowResize: UseWindowResize = (listener, options) => {
+export const useWindowResize: UseWindowResize = (listener, options) => {
   return useWindowEvent('resize', listener, options);
 };
-
-export default useWindowResize;
